fix: new fetch client
